Fix DOWNLOAD response: use %u instead of unsupported %llu
Zephyr's minimal printf on nRF5340 does not support %llu, causing
"bytes":lu in the JSON response. This broke Python JSON parsing and
caused BLE disconnect during recording.

+++ applications/clip/src/at_commands.c
@@ -1393,14 +1393,14 @@ static int cmd_download_handler(struct at_cmd_ctx *ctx, char *response, size_t l
     char data[256];
     if (filename) {
         snprintf(data, sizeof(data),
-                 "{\"state\":\"transmitting\",\"session\":\"%s\",\"file\":\"%s\",\"total\":%u,\"bytes\":%llu}",
+                 "{\"state\":\"transmitting\",\"session\":\"%s\",\"file\":\"%s\",\"total\":%u,\"bytes\":%u}",
                  info.session_id, info.current_file, (unsigned int)info.total_files,
-                 (unsigned long long)info.total_bytes);
+                 (unsigned int)info.total_bytes);
     } else {
         snprintf(data, sizeof(data),
-                 "{\"state\":\"transmitting\",\"session\":\"%s\",\"total\":%u,\"bytes\":%llu}",
+                 "{\"state\":\"transmitting\",\"session\":\"%s\",\"total\":%u,\"bytes\":%u}",
                  info.session_id, (unsigned int)info.total_files,
-                 (unsigned long long)info.total_bytes);
+                 (unsigned int)info.total_bytes);
     }
 
     return create_json_response(true, NULL, data, response, len);
